Add the ability to call members during a meeting and add in-meeting chat functionality.

+++ room/tuiroomkit/src/main/java/com/trtc/uikit/roomkit/RoomHomeActivity.kt
@@ -3,7 +3,7 @@ package com.trtc.uikit.roomkit
 import android.os.Bundle
 import androidx.activity.result.contract.ActivityResultContracts
 import com.trtc.uikit.roomkit.view.RoomHomeView
-import com.trtc.uikit.roomkit.view.RoomScheduleView
+import com.trtc.uikit.roomkit.view.schedule.RoomScheduleView
 import com.trtc.uikit.roomkit.view.schedule.RoomScheduleInfoDialog
 import io.trtc.tuikit.atomicx.common.FullScreenActivity
 import io.trtc.tuikit.atomicxcore.api.room.RoomInfo
@@ -30,8 +30,8 @@ class RoomHomeActivity : FullScreenActivity() {
             roomID = roomID,
             roomName = data.getStringExtra(RoomScheduleView.EXTRA_SCHEDULED_ROOM_NAME).orEmpty(),
             password = data.getStringExtra(RoomScheduleView.EXTRA_SCHEDULED_PASSWORD).orEmpty(),
-            scheduledStartTime = data.getLongExtra(RoomScheduleView.EXTRA_SCHEDULED_START_TIME, 0L),
-            scheduledEndTime = data.getLongExtra(RoomScheduleView.EXTRA_SCHEDULED_END_TIME, 0L),
+            scheduledStartTime = data.getIntExtra(RoomScheduleView.EXTRA_SCHEDULED_START_TIME, 0),
+            scheduledEndTime = data.getIntExtra(RoomScheduleView.EXTRA_SCHEDULED_END_TIME, 0),
         )
         scheduleInfoDialog.show(info)
     }
